+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- [PR-54](https://github.com/itk-dev/aapodwalk/pull/54)
+  - Add scroll into view hook
+  - Scroll active/next point in list into view
 - [PR-52](https://github.com/itk-dev/aapodwalk/pull/52)
   - Avoid zero indexing pins
 - [PR-51](https://github.com/itk-dev/aapodwalk/pull/51)